Switching DOMChange/DOMChanges inner to outer class
