+++ medusa-sample-implementation/src/main/java/com/sample/medusa/eventhandler/Page2EventHandler.java
@@ -2,7 +2,7 @@ package com.sample.medusa.eventhandler;
 
 import io.getmedusa.medusa.core.annotation.PageSetup;
 import io.getmedusa.medusa.core.annotation.UIEventController;
-import io.getmedusa.medusa.core.injector.DOMChange;
+import io.getmedusa.medusa.core.injector.DOMChanges;
 import org.springframework.stereotype.Component;
 
 import java.util.*;
@@ -21,8 +21,8 @@ public class Page2EventHandler implements UIEventController {
     }
 
     //second method with same signature to check if page switching works correctly
-    public DOMChange.DOMChanges buy(Object... parameters) {
-        return DOMChange.of("example-value", "this store is closed!");
+    public DOMChanges buy(Object... parameters) {
+        return DOMChanges.of("example-value", "this store is closed!");
     }
 
 }
